benchmarks/image_processing/requirements.txt
boto3
pillow
pyyaml==6.0.1
pytz==2024.1